* add commentary * use UINTs instead of WORDs to represent a command ID   (this is in line with handlers called by the ON_COMMAND_EX macro)
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/branches/RefactorRevGraph@13666 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/RevisionGraph/RevisionGraphOptions.cpp
@@ -31,7 +31,7 @@ void CRevisionGraphOptionList::Add (IRevisionGraphOption* option)
 
 // utility method
 
-IRevisionGraphOption* CRevisionGraphOptionList::GetOptionByID (WORD id) const
+IRevisionGraphOption* CRevisionGraphOptionList::GetOptionByID (UINT id) const
 {
     for (size_t i = 0, count = options.size(); i < count; ++i)
         if (options[i]->CommandID() == id)
@@ -54,17 +54,17 @@ CRevisionGraphOptionList::~CRevisionGraphOptionList()
 
 // menu interaction
 
-bool CRevisionGraphOptionList::IsAvailable (WORD id) const
+bool CRevisionGraphOptionList::IsAvailable (UINT id) const
 {
     return GetOptionByID (id)->IsAvailable();
 }
 
-bool CRevisionGraphOptionList::IsSelected (WORD id) const
+bool CRevisionGraphOptionList::IsSelected (UINT id) const
 {
     return GetOptionByID (id)->IsSelected();
 }
 
-void CRevisionGraphOptionList::ToggleSelection (WORD id)
+void CRevisionGraphOptionList::ToggleSelection (UINT id)
 {
     GetOptionByID (id)->ToggleSelection();
 }
